refactor: Move session middleware to its own module
